# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "TG Summarizer", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/telegram/scrape": {
      "post": {
        "tags": ["telegram"],
        "summary": "Api Scrape",
        "operationId": "telegram-api_scrape",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScrapeRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Telegram-Api Scrape"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/telegram/channel-info": {
      "post": {
        "tags": ["telegram"],
        "summary": "Api Channel Info",
        "operationId": "telegram-api_channel_info",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelInfoRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Telegram-Api Channel Info"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/telegram/resolve-start-time": {
      "post": {
        "tags": ["telegram"],
        "summary": "Api Resolve Start Time",
        "operationId": "telegram-api_resolve_start_time",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveStartTimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Telegram-Api Resolve Start Time"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/telegram/bot-info": {
      "post": {
        "tags": ["telegram"],
        "summary": "Api Bot Info",
        "operationId": "telegram-api_bot_info",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BotInfoRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Telegram-Api Bot Info"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/telegram/publish": {
      "post": {
        "tags": ["telegram"],
        "summary": "Api Publish",
        "operationId": "telegram-api_publish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PublishRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Telegram-Api Publish"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/telegram/bot-file/{credential_id}": {
      "get": {
        "tags": ["telegram"],
        "summary": "Api Bot File",
        "operationId": "telegram-api_bot_file",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Credential Id" }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "title": "Path" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BotInfoRequest": {
        "properties": {
          "proxyEnabled": {
            "type": "boolean",
            "title": "Proxyenabled",
            "default": false
          },
          "proxies": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Proxies"
          },
          "torAutoRotate": {
            "type": "boolean",
            "title": "Torautorotate",
            "default": false
          },
          "torRotationThreshold": {
            "type": "integer",
            "title": "Torrotationthreshold",
            "default": 10
          },
          "credentialId": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Credentialid"
          },
          "token": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Token"
          },
          "method": { "type": "string", "title": "Method" },
          "params": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Params"
          }
        },
        "type": "object",
        "required": ["method"],
        "title": "BotInfoRequest"
      },
...
      "ChannelInfoRequest": {
        "properties": {
          "proxyEnabled": {
            "type": "boolean",
            "title": "Proxyenabled",
            "default": false
          },
          "proxies": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Proxies"
          },
          "torAutoRotate": {
            "type": "boolean",
            "title": "Torautorotate",
            "default": false
          },
          "torRotationThreshold": {
            "type": "integer",
            "title": "Torrotationthreshold",
            "default": 10
          },
          "channelName": { "type": "string", "title": "Channelname" }
        },
        "type": "object",
        "required": ["channelName"],
        "title": "ChannelInfoRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublishRequest": {
        "properties": {
          "proxyEnabled": {
            "type": "boolean",
            "title": "Proxyenabled",
            "default": false
          },
          "proxies": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Proxies"
          },
          "torAutoRotate": {
            "type": "boolean",
            "title": "Torautorotate",
            "default": false
          },
          "torRotationThreshold": {
            "type": "integer",
            "title": "Torrotationthreshold",
            "default": 10
          },
          "credentialId": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Credentialid"
          },
          "token": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Token"
          },
          "chatId": { "type": "string", "title": "Chatid" },
          "text": { "type": "string", "title": "Text" },
          "metadataText": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Metadatatext"
          }
        },
        "type": "object",
        "required": ["chatId", "text"],
        "title": "PublishRequest"
      },
...
      "ResolveStartTimeRequest": {
        "properties": {
          "proxyEnabled": {
            "type": "boolean",
            "title": "Proxyenabled",
            "default": false
          },
          "proxies": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Proxies"
          },
          "torAutoRotate": {
            "type": "boolean",
            "title": "Torautorotate",
            "default": false
          },
          "torRotationThreshold": {
            "type": "integer",
            "title": "Torrotationthreshold",
            "default": 10
          },
          "channelName": { "type": "string", "title": "Channelname" },
          "targetTimeMs": { "type": "integer", "title": "Targettimems" }
        },
        "type": "object",
        "required": ["channelName", "targetTimeMs"],
        "title": "ResolveStartTimeRequest"
      },
...
      "ScrapeRequest": {
        "properties": {
          "proxyEnabled": {
            "type": "boolean",
            "title": "Proxyenabled",
            "default": false
          },
          "proxies": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Proxies"
          },
          "torAutoRotate": {
            "type": "boolean",
            "title": "Torautorotate",
            "default": false
          },
          "torRotationThreshold": {
            "type": "integer",
            "title": "Torrotationthreshold",
            "default": 10
          },
          "url": { "type": "string", "title": "Url" },
          "knownLatestId": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Knownlatestid"
          },
          "knownDisplayName": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Knowndisplayname"
          },
          "knownPhotoUrl": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Knownphotourl"
          }
        },
        "type": "object",
        "required": ["url"],
        "title": "ScrapeRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}